ref(bigquery): Move more code to a new schema module
