+++ crates/etl-destinations/src/bigquery/mod.rs
@@ -2,6 +2,7 @@ mod client;
 mod core;
 mod encoding;
 mod metrics;
+mod schema;
 mod sql;
 #[cfg(feature = "test-utils")]
 pub mod test_utils;
